Split test_datastore.py into five files by subject area
Pure code motion, verified the same way as the butler split. The file keeps
its history as test_datastore_file.py, which holds the bulk of it.
