+++ tests/test_datastore_file.py
@@ -33,11 +33,8 @@ import contextlib
 import dataclasses
 import logging
 import os
-import pathlib
-import pickle
 import shutil
 import tempfile
-import time
 import unittest.mock
 import uuid
 from collections.abc import Callable, Iterator
@@ -49,7 +46,6 @@ from butler_test_support import records_from
 
 import lsst.daf.butler.datastores.fileDatastore
 from lsst.daf.butler import (
-    Config,
     DataCoordinate,
     DatasetIdGenEnum,
     DatasetRef,
@@ -61,23 +57,7 @@ from lsst.daf.butler import (
     StorageClass,
     StorageClassFactory,
 )
-from lsst.daf.butler.datastore import DatasetRefURIs, DatastoreConfig, DatastoreValidationError, NullDatastore
-from lsst.daf.butler.datastore.cache_manager import (
-    DatastoreCacheManager,
-    DatastoreCacheManagerConfig,
-    DatastoreDisabledCacheManager,
-)
-from lsst.daf.butler.datastore.record_data import (
-    DatastoreRecordData,
-    DatastoreRecordTable,
-    SerializedDatastoreRecordData,
-)
-from lsst.daf.butler.datastore.stored_file_info import (
-    StoredFileInfo,
-    StoredFileInfoTable,
-    make_datastore_path_relative,
-)
-from lsst.daf.butler.datastores.chainedDatastore import ChainedDatastore
+from lsst.daf.butler.datastore import DatastoreConfig, DatastoreValidationError
 from lsst.daf.butler.formatters.yaml import YamlFormatter
 from lsst.daf.butler.tests import (
     BadNoWriteFormatter,
@@ -90,9 +70,7 @@ from lsst.daf.butler.tests import (
 )
 from lsst.daf.butler.tests.dict_convertible_model import DictConvertibleModel
 from lsst.daf.butler.tests.fixtures import make_example_metrics
-from lsst.resources import ResourcePath
 from lsst.utils import doImport
-from lsst.utils.introspection import get_full_type_name
 
 TESTDIR = os.path.dirname(__file__)
 
@@ -1662,1074 +1640,3 @@ def test_cleanup(ds: DatastoreHarness) -> None:
     assert expectedUri.exists(), f"Check for existence of {expectedUri}"
     datastore.remove(ref)
     assert not expectedUri.exists(), f"Check for existence of now removed {expectedUri}"
-
-
-DATA_ID = {
-    "visit": 52,
-    "physical_filter": "V",
-    "band": "v",
-    "instrument": "DummyCamComp",
-    "day_obs": 20250101,
-}
-"""Data ID the constraint configurations are written against."""
-
-OTHER_DATA_ID = {
-    "visit": 52,
-    "physical_filter": "V",
-    "band": "v",
-    "instrument": "HSC",
-    "day_obs": 20250101,
-}
-"""Data ID for a second instrument, which the per-store constraints treat
-differently."""
-
-CONSTRAINT_DATASTORES = [
-    pytest.param("posixDatastoreP.yaml", True, True, id="posix"),
-    pytest.param("inMemoryDatastoreP.yaml", False, False, id="in-memory"),
-    pytest.param("chainedDatastoreP.yaml", True, True, id="chained"),
-    pytest.param("chainedDatastorePa.yaml", True, True, id="chained-native"),
-]
-"""(config file, can ingest, needs a root) for each datastore configuration
-that shares the same constraints.
-
-``chained-memory`` was measured by DM-55822 as zero unique lines and zero
-unique arcs and is gone. ``chained`` measured zero as well, but only because
-``chained-memory`` held the same coverage: between them they were the sole
-cover for `ChainedDatastore.put`'s "child rejects the ref, skip it" branch,
-`chainedDatastore.py` arc 470 to 471. Removing both lost it, so the file-backed
-one stays. A per-axis marginal query cannot see coverage two axes hold jointly.
-"""
-
-CONSTRAINT_CASES = [
-    pytest.param("metric", "StructuredData", True, id="metric"),
-    pytest.param("metric5", "StructuredData", False, id="metric5"),
-    pytest.param("metric33", "StructuredData", True, id="metric33"),
-    pytest.param("metric5", "StructuredDataJson", True, id="metric5-json"),
-]
-"""(dataset type name, storage class, whether the constraint accepts it).
-
-These were subtests of one method; `parametrize` could not be applied while
-they lived on a `unittest.TestCase`.
-"""
-
-PER_STORE_CASES = [
-    pytest.param("metric", DATA_ID, "StructuredData", (False, True, False), True, id="metric"),
-    pytest.param("metric5", DATA_ID, "StructuredData", (False, False, False), False, id="metric5"),
-    pytest.param("metric5", OTHER_DATA_ID, "StructuredData", (True, False, False), False, id="metric5-hsc"),
-    pytest.param("metric33", OTHER_DATA_ID, "StructuredDataJson", (True, True, False), True, id="metric33"),
-    pytest.param("metric5", DATA_ID, "StructuredDataJson", (False, True, False), True, id="metric5-json"),
-]
-"""(dataset type name, data ID, storage class, per-child acceptance, whether
-ingest is expected to work) for the per-store constraint chain."""
-
-
-@pytest.fixture(scope="module")
-def constraint_storage_class_factory() -> StorageClassFactory:
-    """Storage classes for the constraint tests.
-
-    Named distinctly from the plugin's ``storage_class_factory`` because this
-    loads ``storageClasses.yaml`` rather than the Butler configs, matching what
-    ``DatastoreTestsBase.setUpClass`` did. `StorageClassFactory` is a
-    singleton, so the two accumulate rather than conflict.
-    """
-    factory = StorageClassFactory()
-    factory.addFromConfig(os.path.join(TESTDIR, "config/basic/storageClasses.yaml"))
-    return factory
-
-
-def _make_datastore(config_file: str, root: str | None) -> Datastore:
-    """Build a datastore from a test configuration, as the base class did."""
-    path = os.path.join(TESTDIR, "config/basic", config_file)
-    config = DatastoreConfig(path)
-    datastore_type = cast(type[Datastore], doImport(config["cls"]))
-    if root is not None:
-        datastore_type.setConfigRoot(root, config, config.copy())
-    registry = DummyRegistry()
-    return Datastore.fromConfig(config=config.copy(), bridgeManager=registry.getDatastoreBridgeManager())
-
-
-@pytest.fixture
-def testfiles() -> Iterator[dict[str, str]]:
-    """Empty JSON and YAML files, suitable for the ingest checks."""
-    with (
-        tempfile.NamedTemporaryFile(suffix=".yaml") as yaml_file,
-        tempfile.NamedTemporaryFile(suffix=".json") as json_file,
-    ):
-        yield {"yaml": yaml_file.name, "json": json_file.name}
-
-
-def _testfile_for(testfiles: dict[str, str], storage_class_name: str) -> str:
-    """Choose the temporary file whose suffix matches the storage class."""
-    return testfiles["json"] if storage_class_name.endswith("Json") else testfiles["yaml"]
-
-
-@pytest.mark.parametrize(("config_file", "can_ingest", "needs_root"), CONSTRAINT_DATASTORES)
-@pytest.mark.parametrize(("dataset_type_name", "storage_class_name", "accepted"), CONSTRAINT_CASES)
-def test_constraints(
-    config_file: str,
-    can_ingest: bool,
-    needs_root: bool,
-    dataset_type_name: str,
-    storage_class_name: str,
-    accepted: bool,
-    constraint_storage_class_factory: StorageClassFactory,
-    testfiles: dict[str, str],
-    tmp_path_factory: pytest.TempPathFactory,
-) -> None:
-    """Test the constraints model shared by all these datastores."""
-    helper = DatasetTestHelper()
-    root = str(tmp_path_factory.mktemp("datastore")) if needs_root else None
-    datastore = _make_datastore(config_file, root)
-
-    storage_class = constraint_storage_class_factory.getStorageClass(storage_class_name)
-    dimensions = DimensionUniverse().conform(("visit", "physical_filter", "instrument"))
-    testfile = _testfile_for(testfiles, storage_class.name)
-
-    metrics = make_example_metrics()
-    ref = helper.makeDatasetRef(dataset_type_name, dimensions, storage_class, DATA_ID)
-    if accepted:
-        datastore.put(metrics, ref)
-        assert datastore.exists(ref)
-        datastore.remove(ref)
-
-        # Try ingest
-        if can_ingest:
-            datastore.ingest(FileDataset(testfile, [ref]), transfer="link")
-            assert datastore.exists(ref)
-            datastore.remove(ref)
-    else:
-        with pytest.raises(DatasetTypeNotSupportedError):
-            datastore.put(metrics, ref)
-        assert not datastore.exists(ref)
-
-        # Again with ingest
-        if can_ingest:
-            with pytest.raises(DatasetTypeNotSupportedError):
-                datastore.ingest(FileDataset(testfile, [ref]), transfer="link")
-            assert not datastore.exists(ref)
-
-
-@pytest.mark.parametrize(
-    ("dataset_type_name", "data_id", "storage_class_name", "accept", "ingest"), PER_STORE_CASES
-)
-def test_per_store_constraints(
-    dataset_type_name: str,
-    data_id: dict[str, object],
-    storage_class_name: str,
-    accept: tuple[bool, bool, bool],
-    ingest: bool,
-    constraint_storage_class_factory: StorageClassFactory,
-    testfiles: dict[str, str],
-    tmp_path_factory: pytest.TempPathFactory,
-) -> None:
-    """Test that a chained datastore can control constraints per-datastore
-    even if a child datastore would accept.
-    """
-    helper = DatasetTestHelper()
-    root = str(tmp_path_factory.mktemp("datastore"))
-    datastore = _make_datastore("chainedDatastorePb.yaml", root)
-    assert isinstance(datastore, ChainedDatastore)
-
-    storage_class = constraint_storage_class_factory.getStorageClass(storage_class_name)
-    dimensions = DimensionUniverse().conform(("visit", "physical_filter", "instrument"))
-    testfile = _testfile_for(testfiles, storage_class.name)
-
-    metrics = make_example_metrics()
-    ref = helper.makeDatasetRef(dataset_type_name, dimensions, storage_class, data_id)
-    if not any(accept):
-        with pytest.raises(DatasetTypeNotSupportedError):
-            datastore.put(metrics, ref)
-        assert not datastore.exists(ref)
-
-        # Again with ingest
-        with pytest.raises(DatasetTypeNotSupportedError):
-            datastore.ingest(FileDataset(testfile, [ref]), transfer="link")
-        assert not datastore.exists(ref)
-        return
-
-    datastore.put(metrics, ref)
-    assert datastore.exists(ref)
-
-    # Check each datastore inside the chained datastore
-    for child_datastore, expected in zip(datastore.datastores, accept, strict=True):
-        assert child_datastore.exists(ref) == expected, (
-            f"Testing presence of {ref} in datastore {child_datastore.name}"
-        )
-
-    datastore.remove(ref)
-
-    # Check that ingest works
-    if not ingest:
-        with pytest.raises(DatasetTypeNotSupportedError):
-            datastore.ingest(FileDataset(testfile, [ref]), transfer="link")
-        return
-
-    datastore.ingest(FileDataset(testfile, [ref]), transfer="link")
-    assert datastore.exists(ref)
-
-    # Check each datastore inside the chained datastore
-    for child_datastore, expected in zip(datastore.datastores, accept, strict=True):
-        # Ephemeral datastores means InMemory at the moment and that does not
-        # accept ingest of files.
-        if child_datastore.isEphemeral:
-            expected = False
-        assert child_datastore.exists(ref) == expected, (
-            f"Testing presence of ingested {ref} in datastore {child_datastore.name}"
-        )
-
-    datastore.remove(ref)
-
-
-@dataclasses.dataclass
-class CacheFixtures:
-    """Datasets and files shared by the cache tests."""
-
-    root: str
-    """Directory the test files live in."""
-
-    refs: list[DatasetRef]
-    """Simple refs, one per file in ``files``."""
-
-    files: list[ResourcePath]
-    """Files backing ``refs``."""
-
-    composite_refs: list[DatasetRef]
-    """Composite refs, one per entry in ``comp_refs`` and ``comp_files``."""
-
-    comp_refs: list[list[DatasetRef]]
-    """Component refs for each composite."""
-
-    comp_files: list[list[ResourcePath]]
-    """Files backing ``comp_refs``."""
-
-
-@pytest.fixture(scope="module")
-def universe() -> DimensionUniverse:
-    """Dimension universe shared by every test in this module."""
-    return DimensionUniverse()
-
-
-@pytest.fixture(scope="module")
-def cache_storage_class_factory() -> StorageClassFactory:
-    """Storage classes for the cache tests.
-
-    Named distinctly from the plugin's ``storage_class_factory`` because this
-    loads ``storageClasses.yaml`` rather than the Butler configs, matching what
-    ``DatastoreCacheTestCase.setUpClass`` did. `StorageClassFactory` is a
-    singleton, so the two accumulate rather than conflict.
-    """
-    factory = StorageClassFactory()
-    factory.addFromConfig(os.path.join(TESTDIR, "config/basic/storageClasses.yaml"))
-    return factory
-
-
-@pytest.fixture
-def cache(  # numpydoc ignore=PR01
-    tmp_path: pathlib.Path,
-    universe: DimensionUniverse,
-    cache_storage_class_factory: StorageClassFactory,
-) -> CacheFixtures:
-    """Build the refs and files the cache tests operate on."""
-    helper = DatasetTestHelper()
-    root = str(tmp_path)
-
-    # Create some test dataset refs and associated test files
-    sc = cache_storage_class_factory.getStorageClass("StructuredDataDict")
-    dimensions = universe.conform(("visit", "physical_filter"))
-    dataId = {
-        "instrument": "dummy",
-        "visit": 52,
-        "physical_filter": "V",
-        "band": "v",
-        "day_obs": 20250101,
-    }
-
-    # Create list of refs and list of temporary files
-    n_datasets = 10
-    refs = [helper.makeDatasetRef(f"metric{n}", dimensions, sc, dataId) for n in range(n_datasets)]
-
-    root_uri = ResourcePath(root, forceDirectory=True)
-    files = [root_uri.join(f"file{n}.txt") for n in range(n_datasets)]
-
-    # Create test files.
-    for uri in files:
-        uri.write(b"0123456789")
-
-    # Create some composite refs with component files.
-    sc = cache_storage_class_factory.getStorageClass("StructuredData")
-    composite_refs = [helper.makeDatasetRef(f"composite{n}", dimensions, sc, dataId) for n in range(3)]
-    comp_files = []
-    comp_refs = []
-    for n, ref in enumerate(composite_refs):
-        component_refs = []
-        component_files = []
-        for component in sc.components:
-            component_ref = ref.makeComponentRef(component)
-            file = root_uri.join(f"composite_file-{n}-{component}.txt")
-            component_refs.append(component_ref)
-            component_files.append(file)
-            file.write(b"9876543210")
-
-        comp_files.append(component_files)
-        comp_refs.append(component_refs)
-
-    return CacheFixtures(
-        root=root,
-        refs=refs,
-        files=files,
-        composite_refs=composite_refs,
-        comp_refs=comp_refs,
-        comp_files=comp_files,
-    )
-
-
-def _make_cache_manager(config_str: str, universe: DimensionUniverse) -> DatastoreCacheManager:
-    """Build a cache manager from a YAML fragment."""
-    config = Config.fromYaml(config_str)
-    return DatastoreCacheManager(DatastoreCacheManagerConfig(config), universe=universe)
-
-
-def _expiration_config(mode: str, threshold: int | str) -> str:
-    """Return a cache config using the given expiry mode and threshold."""
-    return f"""
-cached:
-  default: true
-  expiry:
-    mode: {mode}
-    threshold: {threshold}
-  cacheable:
-    unused: true
-    """
-
-
-def _assert_cache(cache_manager: DatastoreCacheManager, cache: CacheFixtures) -> None:
-    """Check the manager caches the first ref and refuses the second."""
-    assert cache_manager.should_be_cached(cache.refs[0])
-    assert not cache_manager.should_be_cached(cache.refs[1])
-
-    uri = cache_manager.move_to_cache(cache.files[0], cache.refs[0])
-    assert isinstance(uri, ResourcePath)
-    assert cache_manager.move_to_cache(cache.files[1], cache.refs[1]) is None
-
-    # Check presence in cache using ref and then using file extension.
-    assert not cache_manager.known_to_cache(cache.refs[1])
-    assert cache_manager.known_to_cache(cache.refs[0])
-    assert not cache_manager.known_to_cache(cache.refs[1], cache.files[1].getExtension())
-    assert cache_manager.known_to_cache(cache.refs[0], cache.files[0].getExtension())
-
-    # Cached file should no longer exist but uncached file should be
-    # unaffected.
-    assert not cache.files[0].exists()
-    assert cache.files[1].exists()
-
-    # Should find this file and it should be within the cache directory.
-    with cache_manager.find_in_cache(cache.refs[0], ".txt") as found:
-        assert found.exists()
-        assert found.relative_to(cache_manager.cache_directory) is not None
-
-    # Should not be able to find these in cache
-    with cache_manager.find_in_cache(cache.refs[0], ".fits") as found:
-        assert found is None
-    with cache_manager.find_in_cache(cache.refs[1], ".fits") as found:
-        assert found is None
-
-
-def _assert_expiration(
-    cache_manager: DatastoreCacheManager,
-    cache: CacheFixtures,
-    n_datasets: int,
-    n_retained: int,
-) -> None:
-    """Insert the datasets and then check the number retained."""
-    for i in range(n_datasets):
-        cached = cache_manager.move_to_cache(cache.files[i], cache.refs[i])
-        assert cached is not None
-
-    assert cache_manager.file_count == n_retained
-
-    # The oldest file should not be in the cache any more.
-    for i in range(n_datasets):
-        with cache_manager.find_in_cache(cache.refs[i], ".txt") as found:
-            if i >= n_datasets - n_retained:
-                assert isinstance(found, ResourcePath)
-            else:
-                assert found is None
-
-
-def test_no_cache_dir(cache, universe) -> None:
-    """Test a cache configured with no root directory."""
-    config_str = """
-cached:
-  root: null
-  cacheable:
-    metric0: true
-        """
-    cache_manager = _make_cache_manager(config_str, universe)
-
-    # Look inside to check we don't have a cache directory
-    assert cache_manager._cache_directory is None
-
-    _assert_cache(cache_manager, cache)
-
-    # Test that the cache directory is marked temporary
-    assert cache_manager.cache_directory.isTemporary
-
-
-def test_no_cache_dir_reversed(cache, universe) -> None:
-    """Use default caching status and metric1 to false"""
-    config_str = """
-cached:
-  root: null
-  default: true
-  cacheable:
-    metric1: false
-        """
-    cache_manager = _make_cache_manager(config_str, universe)
-
-    _assert_cache(cache_manager, cache)
-
-
-def test_envvar_cache_dir(cache, universe) -> None:
-    """Test that the cache directory can come from the environment."""
-    config_str = f"""
-cached:
-  root: '{cache.root}'
-  cacheable:
-    metric0: true
-        """
-
-    root = ResourcePath(cache.root, forceDirectory=True)
-    env_dir = root.join("somewhere", forceDirectory=True)
-    elsewhere = root.join("elsewhere", forceDirectory=True)
-
-    # Environment variable should override the config value.
-    with unittest.mock.patch.dict(os.environ, {"DAF_BUTLER_CACHE_DIRECTORY": env_dir.ospath}):
-        cache_manager = _make_cache_manager(config_str, universe)
-    assert cache_manager.cache_directory == env_dir
-
-    # This environment variable should not override the config value.
-    with unittest.mock.patch.dict(os.environ, {"DAF_BUTLER_CACHE_DIRECTORY_IF_UNSET": env_dir.ospath}):
-        cache_manager = _make_cache_manager(config_str, universe)
-    assert cache_manager.cache_directory == root
-
-    # No default setting.
-    config_str = """
-cached:
-  root: null
-  default: true
-  cacheable:
-    metric1: false
-        """
-    cache_manager = _make_cache_manager(config_str, universe)
-
-    # This environment variable should override the config value.
-    with unittest.mock.patch.dict(os.environ, {"DAF_BUTLER_CACHE_DIRECTORY_IF_UNSET": env_dir.ospath}):
-        cache_manager = _make_cache_manager(config_str, universe)
-    assert cache_manager.cache_directory == env_dir
-
-    # If both environment variables are set the main (not IF_UNSET)
-    # variable should win.
-    with unittest.mock.patch.dict(
-        os.environ,
-        {
-            "DAF_BUTLER_CACHE_DIRECTORY": env_dir.ospath,
-            "DAF_BUTLER_CACHE_DIRECTORY_IF_UNSET": elsewhere.ospath,
-        },
-    ):
-        cache_manager = _make_cache_manager(config_str, universe)
-    assert cache_manager.cache_directory == env_dir
-
-    # Use the API to set the environment variable, making sure that the
-    # variable is reset on exit.
-    with unittest.mock.patch.dict(
-        os.environ,
-        {"DAF_BUTLER_CACHE_DIRECTORY_IF_UNSET": ""},
-    ):
-        defined, cache_dir = DatastoreCacheManager.set_fallback_cache_directory_if_unset()
-        assert defined
-        cache_manager = _make_cache_manager(config_str, universe)
-        assert cache_manager.cache_directory == ResourcePath(cache_dir, forceDirectory=True)
-
-    # Now create the cache manager ahead of time and set the fallback
-    # later.
-    cache_manager = _make_cache_manager(config_str, universe)
-    assert cache_manager._cache_directory is None
-    with unittest.mock.patch.dict(
-        os.environ,
-        {"DAF_BUTLER_CACHE_DIRECTORY_IF_UNSET": ""},
-    ):
-        defined, cache_dir = DatastoreCacheManager.set_fallback_cache_directory_if_unset()
-        assert defined
-        assert cache_manager.cache_directory == ResourcePath(cache_dir, forceDirectory=True)
-
-
-def test_explicit_cache_dir(cache, universe) -> None:
-    """Test a cache configured with an explicit root directory."""
-    config_str = f"""
-cached:
-  root: '{cache.root}'
-  cacheable:
-    metric0: true
-        """
-    cache_manager = _make_cache_manager(config_str, universe)
-
-    # Look inside to check we do have a cache directory.
-    assert cache_manager.cache_directory == ResourcePath(cache.root, forceDirectory=True)
-
-    _assert_cache(cache_manager, cache)
-
-    # Test that the cache directory is not marked temporary
-    assert not cache_manager.cache_directory.isTemporary
-
-
-def test_unexpected_files_in_cache_dir(cache, universe) -> None:
-    """Test for regression of a bug where extraneous files in a cache
-    directory would cause all cache lookups to raise an exception.
-    """
-    config_str = f"""
-cached:
-  root: '{cache.root}'
-  cacheable:
-    metric0: true
-        """
-
-    for filename in ["unexpected.txt", "unexpected", "un_expected", "un_expected.txt"]:
-        unexpected_file = os.path.join(cache.root, filename)
-        with open(unexpected_file, "w") as fh:
-            fh.write("test")
-
-    cache_manager = _make_cache_manager(config_str, universe)
-    cache_manager.scan_cache()
-    _assert_cache(cache_manager, cache)
-
-
-def test_no_cache(cache, universe) -> None:
-    """Test that the disabled cache manager caches nothing."""
-    cache_manager = DatastoreDisabledCacheManager("", universe=universe)
-    # unittest formatted the failure message whether or not the assertion
-    # failed, so this was the only caller of the manager's __str__. A bare
-    # assert only formats its message on failure, so compute it up front.
-    message = f"{cache_manager}"
-    for uri, ref in zip(cache.files, cache.refs, strict=True):
-        assert not cache_manager.should_be_cached(ref)
-        assert cache_manager.move_to_cache(uri, ref) is None
-        assert not cache_manager.known_to_cache(ref)
-        with cache_manager.find_in_cache(ref, ".txt") as found:
-            assert found is None, message
-
-
-def test_cache_expiry_files(cache, universe) -> None:
-    """Test that ``files`` expiry retains the threshold number of files."""
-    threshold = 2  # Keep at least 2 files.
-    mode = "files"
-    config_str = _expiration_config(mode, threshold)
-
-    cache_manager = _make_cache_manager(config_str, universe)
-
-    # Check that an empty cache returns unknown for arbitrary ref
-    assert not cache_manager.known_to_cache(cache.refs[0])
-
-    # Should end with datasets: 2, 3, 4
-    _assert_expiration(cache_manager, cache, 5, threshold + 1)
-    assert f"{mode}={threshold}" in str(cache_manager)
-
-    # Check that we will not expire a file that is actively in use.
-    with cache_manager.find_in_cache(cache.refs[2], ".txt") as found:
-        assert found is not None
-
-        # Trigger cache expiration that should remove the file
-        # we just retrieved. Should now have: 3, 4, 5
-        cached = cache_manager.move_to_cache(cache.files[5], cache.refs[5])
-        assert cached is not None
-
-        # Cache should still report the standard file count.
-        assert cache_manager.file_count == threshold + 1
-
-        # Add additional entry to cache.
-        # Should now have 4, 5, 6
-        cached = cache_manager.move_to_cache(cache.files[6], cache.refs[6])
-        assert cached is not None
-
-        # Is the file still there?
-        assert found.exists()
-
-        # Can we read it?
-        data = found.read()
-        assert len(data) > 0
-
-    # Outside context the file should no longer exist.
-    assert not found.exists()
-
-    # File count should not have changed.
-    assert cache_manager.file_count == threshold + 1
-
-    # Dataset 2 was in the exempt directory but because hardlinks
-    # are used it was deleted from the main cache during cache expiry
-    # above and so should no longer be found.
-    with cache_manager.find_in_cache(cache.refs[2], ".txt") as found:
-        assert found is None
-
-    # And the one stored after it is also gone.
-    with cache_manager.find_in_cache(cache.refs[3], ".txt") as found:
-        assert found is None
-
-    # But dataset 4 is present.
-    with cache_manager.find_in_cache(cache.refs[4], ".txt") as found:
-        assert found is not None
-
-    # Adding a new dataset to the cache should now delete it.
-    cache_manager.move_to_cache(cache.files[7], cache.refs[7])
-
-    with cache_manager.find_in_cache(cache.refs[2], ".txt") as found:
-        assert found is None
-
-
-def test_cache_expiry_datasets(cache, universe) -> None:
-    """Test that ``datasets`` expiry retains the threshold count."""
-    threshold = 2  # Keep 2 datasets.
-    mode = "datasets"
-    config_str = _expiration_config(mode, threshold)
-
-    cache_manager = _make_cache_manager(config_str, universe)
-    _assert_expiration(cache_manager, cache, 5, threshold + 1)
-    assert f"{mode}={threshold}" in str(cache_manager)
-
-
-def test_cache_expiry_datasets_from_disabled(cache, universe) -> None:
-    """Test that the expiry-mode envvar enables a disabled cache."""
-    threshold = 2
-    mode = "datasets"
-    with unittest.mock.patch.dict(
-        os.environ,
-        {"DAF_BUTLER_CACHE_EXPIRATION_MODE": f"{mode}={threshold}"},
-    ):
-        cache_manager = DatastoreCacheManager.create_disabled(universe=DimensionUniverse())
-        _assert_expiration(cache_manager, cache, 5, threshold + 1)
-        assert f"{mode}={threshold}" in str(cache_manager)
-
-
-def test_expiration_mode_override(cache, universe, caplog) -> None:
-    """Test that the envvar overrides the configured expiry mode."""
-    threshold = 2  # Keep 2 datasets.
-    mode = "datasets"
-    config_str = _expiration_config(mode, threshold)
-
-    mode = "size"
-    threshold = 55
-    with unittest.mock.patch.dict(
-        os.environ,
-        {"DAF_BUTLER_CACHE_EXPIRATION_MODE": f"{mode}={threshold}"},
-    ):
-        cache_manager = _make_cache_manager(config_str, universe)
-        _assert_expiration(cache_manager, cache, 10, 6)
-        assert f"{mode}={threshold}" in str(cache_manager)
-
-    # Check we get a warning with unrecognized form.
-    with unittest.mock.patch.dict(
-        os.environ,
-        {"DAF_BUTLER_CACHE_EXPIRATION_MODE": "something"},
-    ):
-        with caplog.at_level(logging.WARNING):
-            _make_cache_manager(config_str, universe)
-        assert "Unrecognized form (something)" in caplog.text
-
-    with unittest.mock.patch.dict(
-        os.environ,
-        {"DAF_BUTLER_CACHE_EXPIRATION_MODE": "something=5"},
-    ):
-        with pytest.raises(ValueError, match="Unrecognized value"):
-            _make_cache_manager(config_str, universe)
-
-
-def test_missing_threshold(universe) -> None:
-    """Test that an empty expiry threshold is rejected."""
-    threshold = ""
-    mode = "datasets"
-    config_str = _expiration_config(mode, threshold)
-
-    with pytest.raises(ValueError, match="Cache expiration threshold"):
-        _make_cache_manager(config_str, universe)
-
-
-def test_cache_expiry_datasets_composite(cache, universe) -> None:
-    """Test that ``datasets`` expiry counts a composite as one dataset."""
-    threshold = 2  # Keep 2 datasets.
-    mode = "datasets"
-    config_str = _expiration_config(mode, threshold)
-
-    cache_manager = _make_cache_manager(config_str, universe)
-
-    n_datasets = 3
-    for i in range(n_datasets):
-        for component_file, component_ref in zip(cache.comp_files[i], cache.comp_refs[i], strict=True):
-            cached = cache_manager.move_to_cache(component_file, component_ref)
-            assert cached is not None
-            assert cache_manager.known_to_cache(component_ref)
-            assert cache_manager.known_to_cache(component_ref.makeCompositeRef())
-            assert cache_manager.known_to_cache(component_ref, component_file.getExtension())
-
-    assert cache_manager.file_count == 6  # 2 datasets each of 3 files
-
-    # Write two new non-composite and the number of files should drop.
-    _assert_expiration(cache_manager, cache, 2, 5)
-
-
-def test_cache_expiry_size(cache, universe) -> None:
-    """Test that ``size`` expiry retains files up to the byte threshold."""
-    threshold = 55  # Each file is 10 bytes
-    mode = "size"
-    config_str = _expiration_config(mode, threshold)
-
-    cache_manager = _make_cache_manager(config_str, universe)
-    _assert_expiration(cache_manager, cache, 10, 6)
-    assert f"{mode}={threshold}" in str(cache_manager)
-
-
-def test_disabled_cache(cache, universe) -> None:
-    """Test that the envvar can disable a configured cache."""
-    # Configure an active cache but disable via environment.
-    threshold = 2
-    mode = "datasets"
-    config_str = _expiration_config(mode, threshold)
-
-    with unittest.mock.patch.dict(
-        os.environ,
-        {"DAF_BUTLER_CACHE_EXPIRATION_MODE": "disabled"},
-    ):
-        env_cache_manager = _make_cache_manager(config_str, universe)
-
-    # Configure to be disabled
-    threshold = 0
-    mode = "disabled"
-    config_str = _expiration_config(mode, threshold)
-    cfg_cache_manager = _make_cache_manager(config_str, universe)
-
-    for cache_manager in (
-        cfg_cache_manager,
-        env_cache_manager,
-        DatastoreCacheManager.create_disabled(universe=DimensionUniverse()),
-    ):
-        for uri, ref in zip(cache.files, cache.refs, strict=True):
-            assert not cache_manager.should_be_cached(ref)
-            assert cache_manager.move_to_cache(uri, ref) is None
-            assert not cache_manager.known_to_cache(ref)
-            with cache_manager.find_in_cache(ref, ".txt") as found:
-                assert found is None, f"{cache_manager}"
-            assert "disabled" in str(cache_manager)
-
-
-def test_cache_expiry_age(cache, universe) -> None:
-    """Test that ``age`` expiry removes files older than the threshold."""
-    threshold = 1  # Expire older than 2 seconds
-    mode = "age"
-    config_str = _expiration_config(mode, threshold)
-
-    cache_manager = _make_cache_manager(config_str, universe)
-    assert f"{mode}={threshold}" in str(cache_manager)
-
-    # Insert 3 files, then sleep, then insert more.
-    for i in range(2):
-        cached = cache_manager.move_to_cache(cache.files[i], cache.refs[i])
-        assert cached is not None
-    time.sleep(2.0)
-    for j in range(4):
-        i = 2 + j  # Continue the counting
-        cached = cache_manager.move_to_cache(cache.files[i], cache.refs[i])
-        assert cached is not None
-
-    # Only the files written after the sleep should exist.
-    assert cache_manager.file_count == 4
-    with cache_manager.find_in_cache(cache.refs[1], ".txt") as found:
-        assert found is None
-    with cache_manager.find_in_cache(cache.refs[2], ".txt") as found:
-        assert isinstance(found, ResourcePath)
-
-
-def test_basics() -> None:
-    helper = DatasetTestHelper()
-    storage_class = StorageClassFactory().getStorageClass("StructuredDataDict")
-    ref = helper.makeDatasetRef("metric", DimensionUniverse().empty, storage_class, {})
-
-    null = NullDatastore(None, None)
-
-    assert not null.exists(ref)
-    assert not null.knows(ref)
-    knows = null.knows_these([ref])
-    assert not knows[ref]
-    null.validateConfiguration([ref])
-
-    with pytest.raises(FileNotFoundError):
-        null.get(ref)
-    with pytest.raises(NotImplementedError):
-        null.put("", ref)
-    with pytest.raises(FileNotFoundError):
-        null.getURI(ref)
-    with pytest.raises(FileNotFoundError):
-        null.getURIs(ref)
-    with pytest.raises(FileNotFoundError):
-        null.getManyURIs([ref])
-    with pytest.raises(NotImplementedError):
-        null.getLookupKeys()
-    with pytest.raises(NotImplementedError):
-        null.import_records({})
-    with pytest.raises(NotImplementedError):
-        null.export_records([])
-    with pytest.raises(NotImplementedError):
-        null.export_predicted_records([])
-    with pytest.raises(NotImplementedError):
-        null.export([ref])
-    with pytest.raises(NotImplementedError):
-        null.transfer(null, ref)
-    with pytest.raises(NotImplementedError):
-        null.emptyTrash()
-    with pytest.raises(NotImplementedError):
-        null.trash(ref)
-    with pytest.raises(NotImplementedError):
-        null.forget([ref])
-    with pytest.raises(NotImplementedError):
-        null.remove(ref)
-    with pytest.raises(NotImplementedError):
-        null.retrieveArtifacts([ref], ResourcePath("."))
-    with pytest.raises(NotImplementedError):
-        null.transfer_from({}, [ref])
-    with pytest.raises(NotImplementedError):
-        null.ingest()
-
-
-def test_sequence_access() -> None:
-    """Verify that DatasetRefURIs can be treated like a two-item tuple."""
-    uris = DatasetRefURIs()
-
-    assert len(uris) == 2
-    assert uris[0] is None
-    assert uris[1] == {}
-
-    primary_uri = ResourcePath("1/2/3")
-    component_uri = ResourcePath("a/b/c")
-
-    # affirm that DatasetRefURIs does not support MutableSequence functions.
-    # The type: ignore comments are the point of the test: the assignments are
-    # not allowed, statically or at run time.
-    with pytest.raises(TypeError):
-        uris[0] = primary_uri  # type: ignore[index]
-    with pytest.raises(TypeError):
-        uris[1] = {"foo": component_uri}  # type: ignore[index]
-
-    # but DatasetRefURIs can be set by property name:
-    uris.primaryURI = primary_uri
-    uris.componentURIs = {"foo": component_uri}
-    assert uris.primaryURI == primary_uri
-    assert uris[0] == primary_uri
-
-    primary, components = uris
-    assert primary == primary_uri
-    assert components == {"foo": component_uri}
-
-
-def test_repr() -> None:
-    """Verify __repr__ output."""
-    uris = DatasetRefURIs(ResourcePath("/1/2/3"), {"comp": ResourcePath("/a/b/c")})
-    assert (
-        repr(uris)
-        == 'DatasetRefURIs(ResourcePath("file:///1/2/3"), {\'comp\': ResourcePath("file:///a/b/c")})'
-    )
-
-
-def test_stored_file_info() -> None:
-    helper = DatasetTestHelper()
-    storage_class = StorageClassFactory().getStorageClass("StructuredDataDict")
-    ref = helper.makeDatasetRef("metric", DimensionUniverse().empty, storage_class, {})
-
-    record = dict(
-        storage_class="StructuredDataDict",
-        formatter="lsst.daf.butler.Formatter",
-        path="a/b/c.txt",
-        component="component",
-        checksum=None,
-        file_size=5,
-    )
-    info = StoredFileInfo.from_record(record)
-
-    assert info.to_record() == record
-
-    ref2 = helper.makeDatasetRef("metric", DimensionUniverse().empty, storage_class, {})
-    rebased = info.rebase(ref2)
-    assert rebased.rebase(ref) == info
-
-    with pytest.raises(TypeError):
-        rebased.update(formatter=42)
-
-    with pytest.raises(ValueError, match="Unexpected keyword arguments"):
-        rebased.update(something=42, new="42")
-
-    # Check that pickle works on StoredFileInfo.
-    pickled_info = pickle.dumps(info)
-    unpickled_info = pickle.loads(pickled_info)
-    assert unpickled_info == info
-
-
-def test_make_datastore_path_relative() -> None:
-    assert make_datastore_path_relative("a/relative/path") == "a/relative/path"
-    assert make_datastore_path_relative("path/with#fragment") == "path/with#fragment"
-    assert make_datastore_path_relative("http://server.com/some/path") == "some/path"
-    assert make_datastore_path_relative("http://server.com/some/path#frag") == "some/path#frag"
-
-
-def test_datastore_record_data_json_types() -> None:
-    """Test that we don't round-trip checksums to UUIDs when deserializing
-    datastore record data.
-    """
-    test_json = """
-        {
-            "dataset_ids": [
-                "74478304-abf1-4a9c-9eb2-926090a84446"
-            ],
-            "records": {
-                "lsst.daf.butler.datastore.stored_file_info.StoredFileInfo": {
-                "74478304abf14a9c9eb2926090a84446": {
-                    "file_datastore_records": [
-                    {
-                        "formatter": "lsst.daf.butler.formatters.yaml.YamlFormatter",
-                        "path": "gain_factors/base-2025-158/gain_factors_spx_base-2025-158.yaml",
-                        "storage_class": "GainFactors",
-                        "component": "__NULL_STRING__",
-                        "checksum": "cab515f6-ab67-0484-393f-aaa525dd526f",
-                        "file_size": 5412
-                    }
-                    ]
-                }
-                }
-            }
-        }
-    """
-    id_str = "74478304abf14a9c9eb2926090a84446"
-    s = SerializedDatastoreRecordData.model_validate_json(test_json)
-    assert isinstance(
-        s.records[get_full_type_name(StoredFileInfo)][id_str]["file_datastore_records"][0]["checksum"],
-        str,
-    )
-    dataset_id = uuid.UUID(id_str)
-    d = DatastoreRecordData.from_simple(s)
-    stored = d.records[dataset_id]["file_datastore_records"][0]
-    assert isinstance(stored, StoredFileInfo)
-    assert isinstance(stored.checksum, str)
-
-
-def test_empty_datastore_records_table() -> None:
-    file_info_table = StoredFileInfoTable.from_records([])
-    assert len(file_info_table) == 0
-
-    assert len(DatastoreRecordTable.from_stored_file_info_table("datastore_name", file_info_table)) == 0
-
-    assert len(DatastoreRecordTable.create_empty()) == 0
-    assert len(DatastoreRecordTable.combine([])) == 0
-    # Doesn't throw because there is no mismatch in datastore names.
-    DatastoreRecordTable.create_empty().validate_datastore_names("arbitrary_name")
-
-
-def test_stored_file_info_table_records() -> None:
-    uuid1 = uuid.UUID("019e1892-7b9b-736d-8248-0e031723646c")
-    uuid2 = uuid.UUID("019e1895-9ec3-7431-bed9-8ae60096103f")
-    uuid3 = uuid.UUID("13d13272-454c-4bc4-94d5-3e322982eee8")
-    checksum = (
-        "021ced8799518305c451cde3e921515ef315ee7ba8937"
-        "a92697a20c571c776afa4102744bc28d2d99d35f44e073cde80cf96e387f65f3967cca45b0d015f5a6b"
-    )
-    input_records = [
-        {
-            "dataset_id": uuid1,
-            "path": "a/relative/path.fits",
-            "formatter": "lsst.obs.base.formatters.fitsExposure.FitsExposureFormatter",
-            "storage_class": "ExposureF",
-            "component": "__NULL_STRING__",
-            "checksum": None,
-            "file_size": 123,
-        },
-        {
-            "dataset_id": uuid2,
-            "path": "file:///an/absolute/path.fits",
-            "formatter": "lsst.obs.base.formatters.fitsExposure.FitsExposureFormatter",
-            "storage_class": "ExposureF",
-            "component": "comp",
-            "checksum": checksum,
-            "file_size": -1,
-        },
-    ]
-    table = StoredFileInfoTable.from_records(input_records)
-    assert len(table) == 2
-
-    def _check_records(records: list[dict]) -> None:
-        rec0 = records[0]
-        assert rec0["dataset_id"] == uuid1
-        assert rec0["path"] == "a/relative/path.fits"
-        assert rec0["formatter"] == "lsst.obs.base.formatters.fitsExposure.FitsExposureFormatter"
-        assert rec0["storage_class"] == "ExposureF"
-        assert rec0["component"] is None
-        assert rec0["checksum"] is None
-        assert rec0["file_size"] == 123
-        rec1 = records[1]
-        assert rec1["dataset_id"] == uuid2
-        assert rec1["path"] == "file:///an/absolute/path.fits"
-        assert rec1["formatter"] == "lsst.obs.base.formatters.fitsExposure.FitsExposureFormatter"
-        assert rec1["storage_class"] == "ExposureF"
-        assert rec1["component"] == "comp"
-        assert rec1["checksum"] == checksum
-        assert rec1["file_size"] is None
-
-    arrow_records = table.to_arrow().to_pylist()
-    assert len(arrow_records) == 2
-    _check_records(arrow_records)
-    assert table.to_records() == input_records
-
-    datastore_table = DatastoreRecordTable.from_stored_file_info_table("name_of_datastore", table)
-    datastore_arrow_records = datastore_table.to_arrow().to_pylist()
-    assert len(datastore_arrow_records) == 2
-    _check_records(datastore_arrow_records)
-    assert datastore_arrow_records[0]["datastore_name"] == "name_of_datastore"
-    assert datastore_arrow_records[1]["datastore_name"] == "name_of_datastore"
-    _check_records(datastore_table.to_stored_file_info_table().to_arrow().to_pylist())
-    # Check round-tripping to_arrow() through from_arrow()
-    _check_records(
-        datastore_table.from_arrow(datastore_table.to_arrow())
-        .to_stored_file_info_table()
-        .to_arrow()
-        .to_pylist()
-    )
-
-    with pytest.raises(ValueError, match="do not match known datastores"):
-        datastore_table.validate_datastore_names(["not_the_same_datastore"])
-    datastore_table.validate_datastore_names(["not_the_same_datastore", "name_of_datastore"])
-
-    second_table = DatastoreRecordTable.from_stored_file_info_table(
-        "other_datastore_name",
-        StoredFileInfoTable.from_records(
-            [
-                {
-                    "dataset_id": uuid3,
-                    "path": "a/relative/path2.fits",
-                    "formatter": "lsst.obs.lsst.rawFormatter.LsstCamRawFormatter",
-                    "storage_class": "Exposure",
-                    "component": "__NULL_STRING__",
-                    "checksum": None,
-                    "file_size": 1000,
-                },
-            ]
-        ),
-    )
-    combined_table = DatastoreRecordTable.combine([datastore_table, second_table])
-    combined_records = combined_table.to_arrow().to_pylist()
-    _check_records(combined_records)
-    rec2 = combined_records[2]
-    assert rec2["dataset_id"] == uuid3
-    assert rec2["path"] == "a/relative/path2.fits"
-    assert rec2["formatter"] == "lsst.obs.lsst.rawFormatter.LsstCamRawFormatter"
-    assert rec2["storage_class"] == "Exposure"
-    assert rec2["component"] is None
-    assert rec2["checksum"] is None
-    assert rec2["file_size"] == 1000
-
-    assert len(datastore_table.filter_by_datastore_name("unknown_datastore")) == 0
-    filtered_table = combined_table.filter_by_datastore_name("name_of_datastore")
-    assert len(filtered_table) == 2
-    _check_records(filtered_table.to_arrow().to_pylist())
-    assert filtered_table.to_stored_file_info_table().to_records() == input_records
